Change Qty into float
add "," after ugs

+++ db_medical_store.sql
@@ -197,8 +197,8 @@ create table t_approvisionnement
 	code_produit nvarchar(50),
 	code_fournisseur nvarchar(50),
 	code_depot nvarchar(50),
-    ugs nvarchar ----unite de gestion de stock, milligrammes
-	quantite int,
+    ugs nvarchar, ----unite de gestion de stock, milligrammes
+	quantite float,
 	cout_total decimal(18,0),
 	
  constraint pk_approvisionnement primary key(code_approvisionnement),
